requirements.txt
streamlit
pandas
gsheetsdb
Pillow